requirements.txt
panoptes-utils[images]
typer
watchdog
